Fix ARM32 SVP pointer casts and hardware_rng zero-length request
- sa_svp_buffer_create.c: Use (uint64_t)(uintptr_t) for void*->uint64_t cast
- sa_svp_buffer_release.c: Use (void*)(uintptr_t) for uint64_t->void* cast
- ta.c: Fix both svp_memory casts with uintptr_t; use local void* for
  ta_sa_svp_buffer_release to avoid uint64_t*->void** type punning
- hardware_rng.c: Early return for zero-length RNG request (read(fd,buf,0)
  returns 0 which was misinterpreted as failure); remove trailing whitespace

+++ reference/src/clientimpl/src/sa_svp_buffer_create.c
@@ -21,6 +21,7 @@
 #include "sa.h"
 #include "ta_client.h"
 #include <stdbool.h>
+#include <stdint.h>
 
 sa_status sa_svp_buffer_create(
         sa_svp_buffer* svp_buffer,
@@ -49,7 +50,9 @@ sa_status sa_svp_buffer_create(
         CREATE_COMMAND(sa_svp_buffer_create_s, svp_buffer_create);
         svp_buffer_create->api_version = API_VERSION;
         svp_buffer_create->svp_buffer = *svp_buffer;
-        svp_buffer_create->svp_memory = (uint64_t) svp_memory;
+        // Use uintptr_t intermediate cast to avoid -Wint-to-pointer-cast on
+        // 32-bit platforms where sizeof(void*) != sizeof(uint64_t).
+        svp_buffer_create->svp_memory = (uint64_t)(uintptr_t) svp_memory;
         svp_buffer_create->size = size;
 
         // clang-format off
